# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quotation PDF extraction and management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List all quotations with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: emailId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      summary: Get a single quotation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateQuotation
      tags: [quotations]
      summary: Update a quotation (edit extracted fields or status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuotationBody"
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      summary: Delete a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Items ────────────────────────────────────────────────────────────────────
  /quotations/{id}/items:
    get:
      operationId: listQuotationItems
      tags: [items]
      summary: List all line items for a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuotationItem"
    post:
      operationId: createItem
      tags: [items]
      summary: Add a new line item to a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Quotation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{id}:
    patch:
      operationId: updateItem
      tags: [items]
      summary: Update a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Search ───────────────────────────────────────────────────────────────────
  /search:
    get:
      operationId: searchQuotations
      tags: [quotations]
      summary: Full-text search across quotations and items by part number or supplier
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  # ── Analytics ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Email ──────────────────────────────────────────────────────────────────
...
    Quotation:
      type: object
      properties:
        id:
          type: integer
        emailId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        supplierEmail:
          type: ["string", "null"]
        quotationNumber:
          type: ["string", "null"]
        quotationDate:
          type: ["string", "null"]
        clientAddress:
          type: ["string", "null"]
        clientContact:
          type: ["string", "null"]
        clientVat:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        deliveryTerms:
          type: ["string", "null"]
        totalAmount:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, reviewed, approved, rejected]
        extractionScore:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        pdfStorageKey:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - status
        - createdAt
        - updatedAt
    QuotationWithItems:
      allOf:
        - $ref: "#/components/schemas/Quotation"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/QuotationItem"
          required:
            - items
    UpdateQuotationBody:
      type: object
      properties:
        supplierName:
          type: ["string", "null"]
        supplierEmail:
          type: ["string", "null"]
        quotationNumber:
          type: ["string", "null"]
        quotationDate:
          type: ["string", "null"]
        clientAddress:
          type: ["string", "null"]
        clientContact:
          type: ["string", "null"]
        clientVat:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        deliveryTerms:
          type: ["string", "null"]
        totalAmount:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, reviewed, approved, rejected]
        notes:
          type: ["string", "null"]

    # ── QuotationItem ──────────────────────────────────────────────────────────
    QuotationItem:
      type: object
      properties:
        id:
          type: integer
        quotationId:
          type: integer
        partNumber:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        quantity:
          type: ["string", "null"]
        unitPrice:
          type: ["string", "null"]
        totalPrice:
          type: ["string", "null"]
        leadTime:
          type: ["string", "null"]
        moq:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - quotationId
        - createdAt
        - updatedAt
    UpdateItemBody:
      type: object
      properties:
        partNumber:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        quantity:
          type: ["string", "null"]
        unitPrice:
          type: ["string", "null"]
        totalPrice:
          type: ["string", "null"]
        leadTime:
          type: ["string", "null"]
        moq:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CreateItemBody:
      type: object
      properties:
        partNumber:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        quantity:
          type: ["string", "null"]
        unitPrice:
          type: ["string", "null"]
        totalPrice:
          type: ["string", "null"]
        leadTime:
          type: ["string", "null"]
        moq:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Search ─────────────────────────────────────────────────────────────────
    SearchResults:
      type: object
      properties:
        quotations:
          type: array
          items:
            $ref: "#/components/schemas/Quotation"
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
        total:
          type: integer
      required:
        - quotations
        - items
        - total

    # ── Analytics ──────────────────────────────────────────────────────────────
...